+++ dialogflow-cx/tests/test_basic_webhook_sample.py
@@ -14,63 +14,20 @@
 
 """Dialogflow CX webhook sample unit tests."""
 
-import os
-import uuid
-from contextlib import ExitStack
 
-import mock
 import pytest
-import test_case_delegator as tcd
 from basic_webhook_sample import BasicWebhookSample
-from common import patch_client
-from google.api_core.operation import Operation
-from google.cloud.dialogflowcx import (
-    Agent,
-    ConversationTurn,
-    Flow,
-    Intent,
-    Page,
-    RunTestCaseResponse,
-    TestCase,
-    TestCaseResult,
-    TestResult,
-    TestRunDifference,
-    Webhook,
-)
-from webhook.main import get_webhook_uri
+from utilities import hermetic_test_cases, run_hermetic_test, yield_sample
 
 
 @pytest.fixture(scope="session")
-def project_id():
-    return os.environ["PROJECT_ID"]
-
-
-@pytest.fixture(scope="session")
-def build_uuid():
-    return os.environ["BUILD_UUID"]
-
-
-@pytest.fixture(scope="session")
-def webhook_uri(project_id, build_uuid):
-    return get_webhook_uri(project_id, build_uuid)
-
-
-@pytest.fixture(scope="session")
-def pytest_session_uuid():
-    return uuid.uuid4()
-
-
-@pytest.fixture(scope="session")
-def webhook_sample(project_id, webhook_uri, pytest_session_uuid):
+def webhook_sample(session_uuid, project_id, webhook_uri):
     sample = BasicWebhookSample(
-        agent_display_name=f"Webhook Agent (test session {pytest_session_uuid})",
+        agent_display_name=f"BasicWebhookSample (test session {session_uuid})",
         project_id=project_id,
         webhook_uri=webhook_uri,
     )
-    sample.initialize()
-    yield sample
-    sample.tear_down()
-    del sample
+    yield_sample(sample)
 
 
 @pytest.mark.integration
@@ -86,99 +43,11 @@ def test_basic_webhook_integration(test_case_display_name, webhook_sample):
 
 
 @pytest.mark.hermetic
-@pytest.mark.parametrize(
-    "differences,test_result,xfail",
-    (
-        ([TestRunDifference(description="XFAIL")], TestResult.FAILED, True),
-        ([TestRunDifference(description="XFAIL")], TestResult.PASSED, True),
-        ([], TestResult.FAILED, True),
-        ([], TestResult.PASSED, False),
-    ),
-)
+@pytest.mark.parametrize("differences,test_result,xfail", hermetic_test_cases)
 def test_basic_webhook_hermetic(differences, test_result, xfail):
     sample = BasicWebhookSample(
         agent_display_name="MOCK_AGENT_DISPLAY_NAME",
         project_id=-1,
         webhook_uri="MOCK_WEBHOOK_URI",
     )
-    sample.TEST_CASES = {
-        "Test Case 0": {
-            "input_text": "MOCK_INPUT_TEXT",
-            "expected_response_text": ["MOCK_EXPECTED_RESPONSE_TEXT"],
-        },
-    }
-    with ExitStack() as stack:
-        patch_client(
-            sample.agent_delegator.client,
-            "create_agent",
-            stack,
-            return_value=Agent(name="MOCK_AGENT_NAME"),
-        )
-        patch_client(
-            sample.webhook_delegator.client,
-            "create_webhook",
-            stack,
-            return_value=Webhook(name="MOCK_WEBHOOK_NAME"),
-        )
-        patch_client(
-            sample.intent_delegator.client,
-            "create_intent",
-            stack,
-            return_value=Intent(name="MOCK_INTENT_NAME"),
-        )
-        patch_client(
-            sample.page_delegator.client,
-            "create_page",
-            stack,
-            return_value=Page(name="MOCK_PAGE_NAME"),
-        )
-        patch_client(
-            sample.start_flow_delegator.client,
-            "get_flow",
-            stack,
-            return_value=Flow(name="MOCK_FLOW_NAME"),
-        )
-        patch_client(sample.start_flow_delegator.client, "update_flow", stack)
-        patch_client(sample.page_delegator.client, "delete_page", stack)
-        patch_client(sample.intent_delegator.client, "delete_intent", stack)
-        patch_client(sample.webhook_delegator.client, "delete_webhook", stack)
-        patch_client(sample.agent_delegator.client, "delete_agent", stack)
-        for test_case_delegator in sample.test_case_delegators.values():
-            patch_client(
-                test_case_delegator.client,
-                "create_test_case",
-                stack,
-                return_value=TestCase(
-                    name="MOCK_TEST_CASE_NAME",
-                    display_name="MOCK_TEST_CASE_DISPLAY_NAME",
-                ),
-            )
-            patch_client(test_case_delegator.client, "batch_delete_test_cases", stack)
-
-            def result():
-                return RunTestCaseResponse(
-                    result=TestCaseResult(
-                        test_result=test_result,
-                        conversation_turns=[
-                            ConversationTurn(
-                                virtual_agent_output=ConversationTurn.VirtualAgentOutput(
-                                    differences=differences
-                                )
-                            )
-                        ],
-                    )
-                )
-
-            lro = mock.create_autospec(Operation, instance=True, spec_set=True)
-            lro.result = result
-            patch_client(
-                test_case_delegator.client, "run_test_case", stack, return_value=lro
-            )
-        sample.initialize()
-        for test_case_delegator in sample.test_case_delegators.values():
-            if xfail:
-                with pytest.raises(tcd.DialogflowTestCaseFailure):
-                    test_case_delegator.run_test_case(wait=0)
-            else:
-                test_case_delegator.run_test_case(wait=0)
-        sample.tear_down()
+    run_hermetic_test(sample, differences, test_result, xfail)
